Support rounting through ReactRouter
