+++ src/NotFound.tsx
@@ -1,0 +1,5 @@
+import { Box } from "@mui/system";
+
+export const NotFound: React.VFC = () => {
+    return <Box>Page Not Found.</Box>;
+};
