+++ src/main.tsx
@@ -2,11 +2,21 @@ import { CssBaseline } from "@mui/material";
 import React from "react";
 import ReactDOM from "react-dom";
 import { App } from "./App";
+import { Posts } from "./Posts";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { NotFound } from "./NotFound";
 
 ReactDOM.render(
     <React.StrictMode>
         <CssBaseline />
-        <App />
+        <BrowserRouter>
+            <Routes>
+                <Route path="/" element={<App />}>
+                    <Route index element={<Posts />} />
+                    <Route path="*" element={<NotFound />} />
+                </Route>
+            </Routes>
+        </BrowserRouter>
     </React.StrictMode>,
     document.getElementById("root")
 );
